Q&A/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
pillow==10.1.0
aiofiles==23.2.1 